refactor: rename Linkbar to Categorybar
