+++ src/components/CategoryBar/CategoryBar.css
